+++ CHANGELOG.md
@@ -10,8 +10,7 @@
   - `message`: `TextMessagePacket`, `BinaryMessagePacket` : `MessagePacket`
   - `upgrade`: `UpgradePacket`
   - `noop`: `NoopPacket`
-- [Work in progress] Define the available connection types in the
-  `ConnectionType` enum.
+- Define the available connection types in the `ConnectionType` enum.
 - [Work in progress] Define the `Server` class:
   - [Work in progress] Define the `ServerConfiguration` class to contain the
     available configuration options for the server.